Hrm, still not reproducing error...
